requirements.txt
flask
flask_sqlalchemy
flask_login
flask_babel
flask-wtf
flask-migrate
requests
reportlab
python-dotenv
pytest
pytest-flask
email-validator